diff: a way for debugging init in Android >= 10
when porting to a new device or needing to handle older devices
it becomes extremely hard to debug the early stages.

Android 10 comes with a new concept which means a

- first stage (ramdisk)
- selinux setup (system.img)
- second stage (system.img)

if you fail in the first stage there is almost nothing you can do
for debugging why. until now.
this patch will give a guide (it MUST be adapted VERY likely to your own
paths for the LED) how to do that regardless:

1. Yellow: init main (might be too fast to see it)
2. Red: early mount with switch system as root
3. Green: 1st stage completely done
4. Purple: selinux stage (started from system)
5. Yellow: main init again and again hard to see here
6. Blue: 2nd stage loaded init.rc, logging started

If you ensure that your fstab.{hardware} is in the root of the ramdisk
with the /cache partition flagged with *first_stage_mount* you will even get
a little logging which can be extended ofc.
Example:
/dev/block/platform/soc.0/f9824900.sdhci/by-name/cache         /cache          ext4    nosuid,nodev,barrier=1,noatime,noauto_da_alloc,errors=continue,journal_async_commit,user_xattr,discard      wait,check,first_stage_mount
Hint:
You cannot use /dev/block/bootdevice here as in that symlink will be created in the second stage first.

Besides that you must *NOT* use DT fstab (dtsi kernel config). So DISABLE ANY
fstab early mount there! That config is DEPRECATED anyways and you have to use the fstab
file from Android 10 on. init will still check (atm of writing) for DT fstab entries
and so it might work atm but that might change somewhen.

Using this debug diff will *not* work when you keep it in DT though (at least not here).

> In Android 9 and lower, devices can specify fstab entries for
> early mounted partitions using device tree overlays (DTOs).
> In Android 10 and higher, devices must specify fstab entries for
> early mounted partitions using an fstab file in the first stage ramdisk.

(source: https://source.android.com/devices/architecture/kernel/mounting-partitions-early#fstab-entries)